Fixed RTCC code, removed printf dependency and streamlined code.
Added RTCC code and fixed what was there.  There is now an rtcc command that allows getting and setting
the current time.  The time is set and read correctly, but the RTCC doesn't run; I'm still working on that.

Replaced a call to vsnprintf with a homebrew version since apparently the printf function takes ~10K of
program memory.  The slimmed down version takes < 200 bytes.