Reworked MIB Code to not send parameter headers with each parameter but rather to send a parameter spec that describes the parameters as a third byte in the bus command packet.  The change is done on pic12. On pic24 the current endpoints need to be updated to use the new way of calling rpcs (cf bus_master_rpc_async in bus_master.c) and loading params.
You don't have to call loadparams() anymore.  Just use plist_set_int16(n, val)
to set the values, fill the buffer if there is one and then call
bus_master_rpc_async with the param_spec *including the length of the buffer*.

This code is untested on pic24 but should work.  Currently testing on
pic12.  Note you need to update the current endpoints for it to compile.

All the parameter headers and structs for int16 and buffer are now gone