Reworked MIB code to be more efficient by restructuring message protocol into a stateless protocol for the slave and a more streamlined protocol for the master.  This leads to a dramaticaly reduced size for the mib12_executive.  Currently clocks in at 976 bytes on pic12lf1822.  This is small enough to allow for ping-pong bootoader updating, meaning that mib12_executive is ready to be preprogrammed onto chips at the factory.
Greatly improved test coverage of MIB12 code including simulated slave calls using a
bit-banged i2c implementation as a gpsim stimulus to verify that everything is working
correctly.  Next step is to implement a MIB responder as a gpsim module in order to allow more
thoughout testing including of the master logic and bus collision events.