Finish integrating MIB DSL into pic12 build process
MIB Commands are now specified using a domain specific language.
These .mib files are include with the c and asm files for each
mib12 application module and part of the scons build process is
to compile them into command_map.asm files according to the
template provided in config/templates/command_map.asm.

The build process is working and just a few additional bugs need
to be worked out.

- gsm_module and pv_module are converted to the new build process
- mibtool is updated to allow for generation of command_map.asm
  files as well as dumping mib blocks from hex files