Fixed bugs in pv_module, added debug mib endpoints
There was a random bit being or'ed in with one of the voltage
measurements.  Unclear why I put it there.  I think it was to
test that the fat32 code was write something other than zeros
but it never got removed and only shows up when measuring low
voltages.

Added mib endpoints for sampling v1, v2, v3, i1 for debugging.