Add checksum calculation to mibtool
mibtool checksum <hex_file> calculates the checksum
of the application module pointed to by hex_file in
the same manner as the mib12_executive so that the programmed
rom can be verified.  The application hw type is pulled from
the mib block so it does not need to be specified.