Add ability to read and write flash from commander
This opens up the ability to test all things flash from the command line.
Also, during the writing of this, I realized that you must program the
serial flash one page at a time without crossing a page boundary.  This is
very important since the writes across a page boundary alias back to the
start of the page.