Fix error branching to bootloaded code.
The bootloader now works correctly but requires a direct branch
to the bootloaded code at 0x200 rather than an intermediate step
where the start vector is copied to 0x100 (a reserved location).

The PIC cannot handle a jump to the reserved location however.  I
do not know why but this is reproducible even by hardcoding such a
jump into the mainboard code.  The current method works fine and
the bootloader is now finished.