Streamline, fix and update pic24 mib code.
- (master) no longer send unneeded write to clear the bus.
- (master) always read the return status and value together
- (slave) always send return status+value on all reads, rather
than alternating
- (master) make sure the bus is idle before starting an rpc

The last change listed above makes the pic24 multimaster compatible.
The pic12 code has been multimaster for a long time but multimaster
support was lacking on pic24.  This is now fixed.  A MIB rpc call
should no longer fail ever now, even if the bus is locked.  The pic24
will simply requeue the command and try again later.

If there are intermittent RPC failures, this is a bug and should be
reported for fixing.