Finish v1 of multi_sensor board
Board is based on the elec_sensor_sd board previously made but with
the following changes:
- removed SD card b/c of space
- added sensors
	+ pulse counting (turbine flow sensor)
	+ 3 voltages from 0-30V
	+ 1 current from 0-15A
	+ external serial connection bidirectional and 5V tolerant
	+ wheatstone bridge sensor connection
	+ differential analog voltage
	+ pressure sensor (SM5470 for venturi sensing)

All of the analog sensors are built around a digitally controllable
instrumentation amplifier AD8557.  It's a pricey part but should
give us a lot of flexibility for retargeting the board in the future
and it lets us avoid needing to do lots of calibration.

Other key feature is that all of the analog sensors can automatically
have their input offset voltage nulled out because 2 muxes let you
switch the input polarity of any input.  The signal will then change
sign but the offset will not, so you can determine the offset voltage
without manual calibration.  The AD8557 can be reprogrammed in the
field to change its gain and output offset to compensate.