Add light sleep option to keep Flash memory hot.
Normally the flash memory is powerd down during sleep to conserve
power but when we are pushing firmware, the wakeup time to bring
the flash back up is ~10 ms per 20 byte push RPC since we have time
to sleep after each one.  This makes pushing slower than it needs
to be.

This commit adds a flag that tells the sleep_handler to not power
down the flash memory if we expect to wake up and use it very soon