Add pulse counting code to multi_sensor_module
Upon reception of an RPC, a flag is set.  task() then checks
for that flag, enables the pulse counter (including powering the
flow sensor if attached) for .5 seconds and counts the number of
pulses.  The count can be read by a separate rpc.  This is implemented
in 'multisensor pulse' to read a single value and 'multisensor pulse -c'
to continually acquire and read values.

Voltage is turned on before the half second starts and turned off after
the half second finishes automatically.