Ensure that each scheduled callback is called only once.
This commit makes the semantics of scheduled_rpc callbacks best-effort
in the following sense.  If the bus is locked when the callback is scheduled,
it is queued as an RPC.  If the bus is still locked the next time the callback
is scheduled to run, another RPC is not generated.  Only once the bus is unlocked
and the RPC is sent is the callback cleared to be resent.

This commit has a glaring flaw:
- there is no way to know which callback has succeeded if there is more than one.
  so the code assumes that the lowest index scheduled callback that was in progress
  succeeded.  This does not matter when there is only one callback but we should
  generalize bus_master_rpc_async to pass an argument to the callback in addition
  to the status so that we can distinguish this case and it will be generally
  useful.