First version of infrastructure for generic PIC simulation
Simulator class is a simulator agnostic interface to a PIC simulator
that lets you control the simulator completely.  It spawns a
separate thread to run the simulator in for UI responsiveness.

Concrete simulators are implemented by subclassing SimWorker. Current
basic implementation exists for sim30 to simulate PIC24 code. This
will be used for pic24 unit testing and then pic12 unit tests will
be ported to use this new interface.