Add asynchronous file reading module
It's surprisingly hard to read from a pipe asynchronously in python
in a cross-platform way.  select() and poll() do not work on Windows
when given a pipe or file object.  The only way that seems to work
on all platforms is spinning up a dedicated thread to block reading
the file and then communicate with that thread using a threadsafe
Queue.

This is not efficient for large reading but is fine for reading a
few bytes.  This is needed for safe subprocess communication with
external simulator programs.