Implement reset delay to allow modules to reset themselves
Adds a 100 ms wait from the time ALARM is asserted to the time the
bus power goes off to allow chips to reset and clear all of their
pins.  This disables any additional circuitry on the bus that may
be power the bus parasitically when main power is cut and prevent
a proper reset sequence.