Make rtcc not fill up task loop during long running tasks
This is a bug fix.  If there is a periodically scheduled task
with interval N and a task takes more than taskloop_size*N seconds,
the rtcc_alarm will add the_alarm_callback 32 times to the taskloop
without it ever running.  This fixes that by not adding a task to
the taskloop if the previous addition has not yet finished runing.