Add profiling and fix giant performance problem
Due to the way the SPI peripheral was initialized, it was not running
at full speed.  Evidently, you must set the SPI clock frequency
before setting the SPIEN bit to enable the module.  Failing to do
this causes the setting to be ignored.  This manifests in reading and
writing from flash taking between 3 and 10 times longer than necessary
and causing performance issues in construct_report as well as making
modtool reflash run very slowly.

Using the new perf module, I tracked this down and fixed it.