Add patch for latest svn HEAD of gpsim to fix it.
There are errors in the current head of gpsim that prevent it from
correctly simulating Enhanced Midrange PICs in certain subtle ways
that break unit tests for MoMo.  This patch applies cleanly to
gpsim svn HEAD and fixes the issues.  A bug report has been submitted
with the patch so hopefully this will be fixed upstream in the future.