Update pic12_executive and gsm_module to use local build settings
- Modify scons to allow local build settings for pic12 modules
- Add settings for gsm_module and pic12_executive
- Fix unit testing to show proper target names

Update gsm_module to target both pic16lf1847 and pic16lf1823

- Update pin locations to be in module_settings.json
- Create port_definitions.h shared file that knows which pins can be made analog or not
so that we can safely try to make a pin analog or digital even if that pin does not support
it on a specific chip but the failure is not important.  Without this we get compilation
errors because the right analog register doesn't work.

Bugfixes for GSM module

- Update mib12 API to be compatible with pic16lf1847 and 16lf1823
- Update gsm_serial to allow for alternate serial port pinout
- Update module_settings.json to specify alternate serial port pinout
on v5 hardware

Update build_settings to remove last vestiges of global module configs

Bug fix on pic12lf1822